Added beginngins of a Path DSL that Routers can use
The Path DSL tries (as with the rest of the library) to be minimal.

We decompose paths into Arrays of strings, and search components into Maps. This then lets us use regular scala matchers on the elements of the array

For composing paths into a location string, we have a simple notation, e.g.
((("this" / "that") ? ("q" -> "value") ## "frag").stringify