Added DiffComponent
DiffComponent is intended to be one of the easiest components to work with. Define a render function that returns a DiffNode (e.g., <.div(contents) ). The top-level node must be a stable element tag (as a node cannot replace its own top-level node) but otherwise it can use the diff approach to update itself.