Added fix to HistoryRouter for IE11 (still supported even though old)
IE11 doesn't trigger onpopstate if the change is only to the hashfragment. Accordingly, if the user agent contains Trident (IE's rendering engine) it also adds a listener for hashchange events.

Although user agent strings are not always reliable, in this case the result of a "miss" is just to evaluate the route twice.