Altered the DNode lifecycle
Previously, DNode.attach called afterAttach() on any children immediately after it had called attach() on them. The advantage of this approach had been that any components delegating to another VNode only needed to forward attach not all the other lifecycle calls.

However, I ran into a problem where it was calling afterAttach() on children before their parent was inserted into the DOM. This would cause some calls on dom elements, such as getBBox(), to fail because the node is not yet in the DOM tree.