Upgrade to Scala 2.13
The main impact comes from Scala changing scala.Seq to be an alias for collection.immutable.Seq. In the differ, and consequently in DiffNodes and DNodes, we'd like to minimise collection-copying but remember that when given a target sequence of VNodes we might replace some of them with VNodes that are already in the tree (so we'd like to be able to mutate the references in the sequence).