Moved layout into the second argument list for VSlides and Sequencer
Case classes perform their equality check on the first argument list to the constructor. As anonymous functions (even if they are the same) are rarely "equal" to each other, this would cause many re-renderings (e.g. changing slide) to cause the VSlides component to completely replace itself. Though this is fast, it would cause embedded iframes (e.g. videos) to reload even if the same slide was clicked on, so I've made this tweak to see if it reduces the likelihood of a slide's embedded video reloading.