Fix bug where items with keys could be duplicated
In the Differ algorithm, if a key on the left was skipped ("is moving down") because it matched a key on the right, but then the list emptied, the right-hand element was being appended. This meant the left element was neither removed nor reused, but the right element was added, duplicating the item.