Markup now offers Fixed and Settable options
The way markup nodes (e.g. for rendering Markdown) worked previously had some awkward side-effects, particularly because in JS anonymous functions are not equal even when equivalent.
In practice, in some sites this caused Markdown nodes to be treated as not being equal even when their content had not changed, and the reconciliation process often meant that nodes after them were removed and re-created too.

This version introduced Fixed (which is tested for equality based on its text) and Settable (which uses MakeItSo to update) and Updatable (which uses update). If your text is dynamic, it's generally encouraged to use Updatable