Fix bug where if a DNode was removed, but one of its children moved elsewhere in the tree, the child would be erroneously detached
Suppose we have a VNode that moves from within one DiffNode to within another and the old parent is removed in the same update
(e.g., this occured when a Scatter tile moved from a "free tile container" into a Socket).

If the new parent was processed first, it would adopt the VNode and move its dom element into the correct place in the tree. The later removal of the old element, however, would then call detach on its children (erroneously including the moved node), causing it to seem to disappear.

The fix ensures that DNode only calls detach on its children if the dom element is still one of its children (i.e. the node has not been already adopted and moved by a new parent)