Scatter's TileComponent trait no longer defaults to using the measured on-screen size as its size implementation
Although it's convenient from a programming perspective just to get the size of a dom element (using getBBox or getClientBoundingRect, etc), it can have performance issues as it requires the browser to do a layout pass.

This commit changes TileComponent so that it no longer defaults size to the measured size at the trait level. However, at the moment, some TileComponent subclasses do use the measured size in their implementation (e.g. TileText and TileForeignObject)