Added Styling and StyleSuite
This is a very simple CSS-in-JS functionality.

A Styling represents a set of rules, alongside a set of rules for modification. For example

val someHead = Styling("text-decoration: underline").modifiedBy(":hover" -> "color: blue").register()

These are contained within a StyleSuite. A StyleSuite's job is to insert a style element into the head of the document, containing the stylings. These use randomly generated css class names.

So, we might see

<style id="akdhfa" type="text/css">
  .ekwrqw {
    text-decoration: underline;
  }

  .ekwrqw:hover {
    color: blue;
  }
</style>

The styling can then be applied to elements by adding the css class, e.g.

<.h1(^.cls := someHead.className, "This should be styled")