Added Unique
Unique (and html.unique, which is an alias for it) helps solve a common mistake:

val myPage = <.div("Some content")

The problem with keeping the div in a val being that during page updates, the div can be asked to morph itself to have different content. This can be confusing when using a doctacular site, because it looks as if navigating between pages hasn't worked (when it has, but the contents of myPage have changed to match another one during a reconciliation process). Previously, we had to remember to do

def myPage = <.div("Some content")

but this update now also allows

val myPage = unique(<.div("Some content"))

The mechanism is that Unique[N] just wraps a delegate VNode[N]. Because Unique[N] does not implement MakeItSo, it cannot be asked to morph. (Though beware that individually, its children can - it doesn't fix the whole tree, though usually that's not necessary).