Fix bug where Stylings were given class names above the BMP
There was an issue with Firefox where stylings randomly would not show. This appears to have been due to having too high a character range in the CSS class names - limiting it to the Basic Multilingual Plane (up to 0xD7FF) appears to fix it and, for instance, avoids the issue of running into the range where surrogate pairs (pairs of 16 bit characters) are used to represent higher characters.