Added Retention and Keep traits
Previously, we've done retention in the default reconciler strictly on the equality of the item. However, that feels hacky at times, e.g. if we have a Morphing component is it really "equal" to the same component with different properties?

With this update, VNodes return a Retention strategy. The default is to try to retain the node if it is equal to the target, however we can now implement Keep(value) if we want to retain the node if some specified values within it are equal.