Introduces Blueprint
MakeItSo items are inherently mutable, as we ask them to morph themselves to match an instruction or destination. However, we often prefer to program using immutable classes.

Blueprint is a trait that can be implemented to denote an immutable blueprint for creating a MakeItSo item. It is intended for use with the Keep and Keyed retention strategies, and the Blueprint also needs to implement those strategies for it to be efficient (otherwise the reconciler will have to build the items multiple times to reach the Keeps/Keys)

Reconcilers accept Blueprints in place of real items in target lists, inflating the Blueprint into an item where necessary.