Added clearer types for DElements etc
We had been overusing V (using it for both VNodes and type aliases for DElements in HTML). And had some cases where for instance VHtmlComponent was actually typed to components producing DElement[dom.Element] rather than DElement[dom.html.Element]

This commit creates a clearer naming scheme

D is for diff / dynamic
V is for VNode stuff

So,

DHtmlComponents produce DHtmlContent, which are DElement[dom.html.Element] or Blueprints for it

VHtmlNode is a VNode[dom.html.Node]

etc.

This commit doesn't yet update the templates, scatter, etc, as I wanted to leave a commit in showing the changes just to core