Introduces stateVariable and Animator
Manually calling rerender is something people don't like to see. This commit introduces stateVariables and the Animator, which will automatically trigger a rerender on a component in the next animation frame.

The Animator is a simple batching of tasks to requestAnimationFrame, with the advantage that as we call all the queued tasks with the same timestamp, we have an easy way to know if a component has already been rerendered in this frame and doesn't need reprocessing.

It is still possible for a component to end up rendering twice or more, as a parent might get rerendered and the cascading call to update doesn't (yet) check the Animator. That's something to think about for later - it might not happen very often in a way that'd be noticeable.