DynamicElements can render dynamic variables containing modifiers
Although this is a fairly basic / unrefined version, it seems to work quite nicely.

DynamicSource becomes covariant to fix a variance issue now they can be passed in as element children.

DynamicValue has a tearDown method added, mostly on behalf of DynamicElement. Suppose a dynamic element has two dynamic values it depends on. One clearing would trigger a resync, but during the call to resync the other might still have a value. This would be annoying if we used "clear" to immediately free it for garbage collection, as its clear listener would trigger a request for another resync next frame. Instead we keep a private dependent value inside the component, and call tearDown.