Update docs for dynamic state and fix bugs
One of the bugs bears further investigation:

At the moment, DynamicElement implements its own machinery to check if it has already rendered in this animation frame. If instead we get it to extend AnimationAdapter, then for some reason something tries to call animationUpdate from DiffComponent even though it is not a related type.